+++ config/locales/sv.yml
@@ -484,7 +484,7 @@ sv:
       edit_branding_status: Redigera Status
       branding_license: Accepterar policy
       pay_branding_fee_tooltip: Betald avgift förlänger H-märkningen i ett år.
-      branding_fee_expire_date_tooltip: Branding-licensen löper ut efter detta datum
+      branding_fee_expire_date_tooltip: Efter detta datum går perioden för din H-märkning ut.
       members: Företagsmedlemmar
 
     index:
@@ -622,7 +622,7 @@ sv:
       submit_new_password: Spara det nya lösenordet
       member: *member
       pay_membership_tooltip: När du betalar din medlemsavgift förlängs ditt medlemskap i ett år.
-      membership_expire_date_tooltip: Medlemskapet löper ut efter detta datum
+      membership_expire_date_tooltip: Efter detta datum går perioden för ditt medlemskap ut.
 
     update:
       passwords_dont_match: De två lösenord du angivit matchar inte.
