+++ features/member_edits_company_page.feature
@@ -8,6 +8,8 @@ Feature: As a member
     Given the following users exists
       | email               | admin | member |
       | emma@happymutts.com |       | true   |
+      | member@random.com   |       | true   |
+      | user@random.com     |       |        |
       | admin@shf.se        | true  | false  |
 
     Given the following payments exist
@@ -118,14 +120,14 @@ Feature: As a member
       | Happy Mutts            | 5562252998                    | kicki@gladajyckar.se | http://www.gladajyckar.se      |
     And I click on t("submit")
     And I am Logged out
-    And I am logged in as "applicant_2@random.com"
+    And I am logged in as "member@random.com"
     And I am on the "edit my company" page for "emma@happymutts.com"
     Then I should be on the "landing" page
     And I should see t("errors.not_permitted")
 
 
   Scenario: User tries to go to company page (gets rerouted)
-    Given I am logged in as "applicant_2@random.com"
+    Given I am logged in as "user@random.com"
     And I am on the "edit my company" page for "emma@happymutts.com"
     Then I should be on the "landing" page
     And I should see t("errors.not_permitted")
