+++ features/shf-documents/view-all-shf-documents.feature
@@ -110,6 +110,6 @@ Feature: SHF members (and admins) can views board meeting minutes (SHF documents
 
 
   Scenario: A user cannot see SHF documents
-    Given I am logged in as "bob@snarkybarky.se "
+    Given I am logged in as "bob@snarkybarky.se"
     And I am on the "all SHF documents" page
     Then I should see t("errors.not_permitted")
